requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
PyPDF2==3.0.1
langchain==0.0.3
langchain-google-genai==0.0.1
google-generativeai==0.0.1
pydantic==2.0.1
faiss-cpu==1.7.4
google-auth==2.22.0
